Fix integraiton test
