+++ tests/integration/test_charm_no_runner.py
@@ -9,8 +9,7 @@ from juju.application import Application
 from juju.model import Model
 
 from charm_state import BASE_VIRTUAL_MACHINES_CONFIG_NAME
-from manager_service import GITHUB_RUNNER_MANAGER_SERVICE_NAME
-from tests.integration.helpers.common import reconcile, run_in_unit, wait_for
+from tests.integration.helpers.common import reconcile, wait_for
 from tests.integration.helpers.openstack import OpenStackInstanceHelper
 
 logger = logging.getLogger(__name__)
@@ -85,34 +84,3 @@ async def test_reconcile_runners(
     await reconcile(app=app, model=model)
 
     await wait_for(lambda: _runners_number(0), timeout=10 * 60, check_interval=10)
-
-
-@pytest.mark.asyncio
-@pytest.mark.abort_on_fail
-async def test_manager_service_started(
-    model: Model,
-    app_no_runner: Application,
-    instance_helper: OpenStackInstanceHelper,
-) -> None:
-    """
-    arrange: A working application with no runners.
-    act: Check the github runner manager service.
-    assert: The service should be running.
-    """
-    app = app_no_runner
-    unit = app.units[0]
-
-    code, stdout, stderr = await run_in_unit(
-        unit,
-        f"sudo systemctl status {GITHUB_RUNNER_MANAGER_SERVICE_NAME}",
-        timeout=60,
-        assert_on_failure=False,
-        assert_msg="GitHub runner manager service not healthy",
-    )
-    print("=====================")
-    print(code)
-    print("=====================")
-    print(stdout)
-    print("=====================")
-    print(stderr)
-    print("=====================")
